updated
git-svn-id: http://svn.codehaus.org/jetty/servlet-api/branches/servlet-3.0@5197 c7fae8f8-2506-0410-b6b2-ad6cf81c2737

+++ src/main/java/javax/servlet/annotation/MultipartConfig.java
@@ -41,10 +41,10 @@ import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
-
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 public @interface MultipartConfig {
+
     /**
      * The directory location where files will be stored
      * 
@@ -52,16 +52,17 @@ public @interface MultipartConfig {
     String location() default "";
 
     /**
-     * the maximum size allowed for files uploaded
+     * The maximum size allowed for files uploaded
      *
      */
-    int maxFileSize() default 0;
+    long maxFileSize() default 0L;
 
     /**
-     * The maximum size of a multi-part/form-data request allowed
+     * The maximum size allowed for <tt>multipart/form-data</tt>
+     * requests
      *
      */
-    int maxRequestSize() default 0;
+    long maxRequestSize() default 0L;
 
     /**
      * The size threshold after which the file will be written to disk
